refactoring: extract data gestion to componant
