+++ types/Components/index.ts
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from "react";
-import { ArtistItems, TrackItems } from "../spotify/index";
+import { ArtistItems, QueryItems, TrackItems } from "../spotify/index";
 import { ArtistData, TrackData } from "../deezer";
 
 export interface ArtistCardProps {
@@ -42,4 +42,8 @@ export interface HamburgerProps {
 export interface PeriodChoiceStyle {
   search__container: string;
   active: string;
+}
+
+export interface ArtistWrapperProps {
+  queryParams: QueryItems;
 }
